refactor(members): put more actions into list item actions
